rename RuleTableMetaDataLoader to RuleMetaDataLoader
