+++ encrypt-core/encrypt-core-common/src/main/resources/META-INF/services/org.apache.shardingsphere.underlying.common.metadata.schema.spi.RuleMetaDataDecorator
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.metadata.ShardingTableMetaDataDecorator
+org.apache.shardingsphere.encrypt.metadata.EncryptMetaDataDecorator
