+++ master-slave-core/master-slave-core-common/src/main/resources/META-INF/services/org.apache.shardingsphere.underlying.common.metadata.schema.spi.RuleMetaDataLoader
@@ -1,3 +1,4 @@
+
 #
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
@@ -15,4 +16,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.encrypt.metadata.EncryptTableMetaDataDecorator
+org.apache.shardingsphere.masterslave.metadata.MasterSlaveMetaDataLoader
