+++ sharding-core/sharding-core-common/src/main/resources/META-INF/services/org.apache.shardingsphere.underlying.common.metadata.schema.spi.RuleMetaDataDecorator
@@ -1,4 +1,3 @@
-
 #
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
@@ -16,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.masterslave.metadata.MasterSlaveTableMetaDataLoader
+org.apache.shardingsphere.core.metadata.ShardingMetaDataDecorator
