+++ sharding-core/sharding-core-common/src/main/resources/META-INF/services/org.apache.shardingsphere.underlying.common.metadata.schema.spi.RuleMetaDataLoader
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.metadata.ShardingTableMetaDataLoader
+org.apache.shardingsphere.core.metadata.ShardingMetaDataLoader
