+++ sharding-proxy/sharding-proxy-backend/src/main/java/org/apache/shardingsphere/shardingproxy/backend/schema/impl/ShardingSchema.java
@@ -21,10 +21,10 @@ import com.google.common.eventbus.Subscribe;
 import lombok.Getter;
 import org.apache.shardingsphere.api.config.sharding.ShardingRuleConfiguration;
 import org.apache.shardingsphere.core.log.ConfigurationLogger;
-import org.apache.shardingsphere.core.metadata.ShardingTableMetaDataDecorator;
+import org.apache.shardingsphere.core.metadata.ShardingMetaDataDecorator;
 import org.apache.shardingsphere.core.rule.MasterSlaveRule;
 import org.apache.shardingsphere.core.rule.ShardingRule;
-import org.apache.shardingsphere.encrypt.metadata.EncryptTableMetaDataDecorator;
+import org.apache.shardingsphere.encrypt.metadata.EncryptMetaDataDecorator;
 import org.apache.shardingsphere.orchestration.core.common.event.ShardingRuleChangedEvent;
 import org.apache.shardingsphere.orchestration.core.common.rule.OrchestrationMasterSlaveRule;
 import org.apache.shardingsphere.orchestration.core.common.rule.OrchestrationShardingRule;
@@ -198,9 +198,9 @@ public final class ShardingSchema extends LogicSchema {
                 LogicSchemas.getInstance().getDatabaseType(), getBackendDataSource().getDataSources(), tableName, ShardingProxyContext.getInstance().getProperties());
         if (tableMetaData.isPresent()) {
             TableMetaData result = tableMetaData.get();
-            result = new ShardingTableMetaDataDecorator().decorate(tableName, result, shardingRule);
+            result = new ShardingMetaDataDecorator().decorate(tableName, result, shardingRule);
             if (!shardingRule.getEncryptRule().getEncryptTableNames().isEmpty()) {
-                result = new EncryptTableMetaDataDecorator().decorate(tableName, result, shardingRule.getEncryptRule());
+                result = new EncryptMetaDataDecorator().decorate(tableName, result, shardingRule.getEncryptRule());
             }
             return Optional.of(result);
         }
