+++ shardingsphere-underlying/shardingsphere-common/src/main/java/org/apache/shardingsphere/underlying/common/metadata/schema/RuleSchemaMetaDataLoader.java
@@ -24,8 +24,8 @@ import org.apache.shardingsphere.sql.parser.binder.metadata.schema.SchemaMetaDat
 import org.apache.shardingsphere.sql.parser.binder.metadata.table.TableMetaData;
 import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationProperties;
 import org.apache.shardingsphere.underlying.common.database.type.DatabaseType;
-import org.apache.shardingsphere.underlying.common.metadata.schema.spi.RuleTableMetaDataDecorator;
-import org.apache.shardingsphere.underlying.common.metadata.schema.spi.RuleTableMetaDataLoader;
+import org.apache.shardingsphere.underlying.common.metadata.schema.spi.RuleMetaDataDecorator;
+import org.apache.shardingsphere.underlying.common.metadata.schema.spi.RuleMetaDataLoader;
 import org.apache.shardingsphere.underlying.common.rule.BaseRule;
 import org.apache.shardingsphere.underlying.common.rule.TablesAggregationRule;
 
@@ -45,8 +45,8 @@ import java.util.TreeSet;
 public final class RuleSchemaMetaDataLoader {
     
     static {
-        ShardingSphereServiceLoader.register(RuleTableMetaDataLoader.class);
-        ShardingSphereServiceLoader.register(RuleTableMetaDataDecorator.class);
+        ShardingSphereServiceLoader.register(RuleMetaDataLoader.class);
+        ShardingSphereServiceLoader.register(RuleMetaDataDecorator.class);
     }
     
     private final Collection<BaseRule> rules;
@@ -64,7 +64,7 @@ public final class RuleSchemaMetaDataLoader {
     public SchemaMetaData load(final DatabaseType databaseType, final Map<String, DataSource> dataSourceMap, final ConfigurationProperties properties) throws SQLException {
         SchemaMetaData result = new SchemaMetaData(new HashMap<>());
         Collection<String> excludedTableNames = new TreeSet<>(String.CASE_INSENSITIVE_ORDER);
-        for (Entry<BaseRule, RuleTableMetaDataLoader> entry : OrderedSPIRegistry.getRegisteredServices(rules, RuleTableMetaDataLoader.class).entrySet()) {
+        for (Entry<BaseRule, RuleMetaDataLoader> entry : OrderedSPIRegistry.getRegisteredServices(rules, RuleMetaDataLoader.class).entrySet()) {
             SchemaMetaData schemaMetaData = entry.getValue().load(databaseType, dataSourceMap, entry.getKey(), properties, excludedTableNames);
             excludedTableNames.addAll(schemaMetaData.getAllTableNames());
             if (entry.getKey() instanceof TablesAggregationRule) {
@@ -104,7 +104,7 @@ public final class RuleSchemaMetaDataLoader {
     @SuppressWarnings("unchecked")
     public Optional<TableMetaData> load(final DatabaseType databaseType, 
                                         final Map<String, DataSource> dataSourceMap, final String tableName, final ConfigurationProperties properties) throws SQLException {
-        for (Entry<BaseRule, RuleTableMetaDataLoader> entry : OrderedSPIRegistry.getRegisteredServices(rules, RuleTableMetaDataLoader.class).entrySet()) {
+        for (Entry<BaseRule, RuleMetaDataLoader> entry : OrderedSPIRegistry.getRegisteredServices(rules, RuleMetaDataLoader.class).entrySet()) {
             Optional<TableMetaData> result = entry.getValue().load(databaseType, dataSourceMap, tableName, entry.getKey(), properties);
             if (result.isPresent()) {
                 return Optional.of(decorate(tableName, result.get()));
@@ -133,9 +133,9 @@ public final class RuleSchemaMetaDataLoader {
     @SuppressWarnings("unchecked")
     private SchemaMetaData decorate(final SchemaMetaData schemaMetaData) {
         Map<String, TableMetaData> result = new HashMap<>(schemaMetaData.getAllTableNames().size(), 1);
-        Map<BaseRule, RuleTableMetaDataDecorator> decorators = OrderedSPIRegistry.getRegisteredServices(rules, RuleTableMetaDataDecorator.class);
+        Map<BaseRule, RuleMetaDataDecorator> decorators = OrderedSPIRegistry.getRegisteredServices(rules, RuleMetaDataDecorator.class);
         for (String each : schemaMetaData.getAllTableNames()) {
-            for (Entry<BaseRule, RuleTableMetaDataDecorator> entry : decorators.entrySet()) {
+            for (Entry<BaseRule, RuleMetaDataDecorator> entry : decorators.entrySet()) {
                 result.put(each, entry.getValue().decorate(each, schemaMetaData.get(each), entry.getKey()));
             }
         }
@@ -145,7 +145,7 @@ public final class RuleSchemaMetaDataLoader {
     @SuppressWarnings("unchecked")
     private TableMetaData decorate(final String tableName, final TableMetaData tableMetaData) {
         TableMetaData result = tableMetaData;
-        for (Entry<BaseRule, RuleTableMetaDataDecorator> entry : OrderedSPIRegistry.getRegisteredServices(rules, RuleTableMetaDataDecorator.class).entrySet()) {
+        for (Entry<BaseRule, RuleMetaDataDecorator> entry : OrderedSPIRegistry.getRegisteredServices(rules, RuleMetaDataDecorator.class).entrySet()) {
             result = entry.getValue().decorate(tableName, tableMetaData, entry.getKey());
         }
         return result;
